refactor(cli): move vendored bazelisk/core.go code to bazelisk-core.go (#7116)
Vendor the bazelisk/core.go file with minimal changes. Some of the
vendored code with a diff is still within bazelisk.go and can maybe be
refactored later

---

### Changes are visible to end-users: no

### Test plan

- Covered by existing test cases
- Manual testing; please provide instructions so we can reproduce: run
the pro cli

GitOrigin-RevId: 0beca33dce43bf7c51c962beef2851ca3a72ad63
